+++ client/css/style.css
@@ -209,27 +209,27 @@
   display: none;
 }
 
-.home-baner {
+.home-banner {
   position: relative;
   cursor: pointer;
   margin: 0 auto;
 }
 
-.home-baner__wrapper {
+.home-banner__wrapper {
   position: absolute;
   margin-left: -960px;
   top: 0;
   left: 50%;
 }
 
-.home-baner__today-img-container {
+.home-banner__today-img-container {
   position: absolute;
   left: 1280px;
   top: 45px;
   height: 360px;
 }
 
-.home-baner__today-img-container li {
+.home-banner__today-img-container li {
   position: relative;
   height: 60px;
   border: 2px none;
