+++ client/js/main/home-banner.js
