+++ client/scss/style.scss
@@ -141,7 +141,7 @@
   display: none;
 }
 
-.home-baner {
+.home-banner {
   position: relative;
   cursor: pointer;
   margin: 0 auto;
@@ -152,7 +152,7 @@
     left: 50%;
   }
 }
-.home-baner__today-img-container {
+.home-banner__today-img-container {
   position: absolute;
   left: 1280px;
   top: 45px;
